# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/reservations/hold": {
      "post": {
        "description": "픽업 시각 검증 → 전체 아이템 재고 확인 → All-or-Nothing 방식으로 Redis Hold 저장 (TTL 2분).\n\n**All-or-Nothing 규칙**\n- 요청한 모든 아이템의 재고가 충분할 때만 holdToken을 발급하고 Redis에 Hold를 생성합니다.\n- 단 하나라도 재고 부족이면 Redis Hold를 생성하지 않고 `success: false`와 상세 실패 목록을 반환합니다.\n\n**세션 상태 전이**\n- 성공: `PRE_HOLD_CONFIRM → WAITING_FOR_CONFIRM` (hold_token 함께 저장)\n- 실패: 현재 상태 → `FAIL` (last_error에 실패 사유 기록)\n\n실제 재고 차감은 `/confirm` 단계에서만 수행됩니다.",
        "operationId": "holdReservation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHoldDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "일부 또는 전체 재고 부족 — holdToken 미발급 (success: false)",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "success": false,
                      "holdToken": null,
                      "items": [
                        {
                          "id": "101",
                          "name": "소금빵",
                          "requestedCount": 2,
                          "heldCount": 2,
                          "status": "SUCCESS"
                        },
                        {
                          "id": "202",
                          "name": "고구마빵",
                          "requestedCount": 3,
                          "heldCount": 0,
                          "status": "OUT_OF_STOCK",
                          "reason": "재고 부족 (남은 수량: 1개)"
                        }
                      ]
                    },
                    "message": "Hold created successfully"
                  }
                }
              }
            }
          },
          "201": {
            "description": "전체 hold 성공 — holdToken 발급",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "success": true,
                      "holdToken": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "items": [
                        {
                          "id": "101",
                          "name": "소금빵",
                          "requestedCount": 2,
                          "heldCount": 2,
                          "status": "SUCCESS"
                        },
                        {
                          "id": "202",
                          "name": "크림빵",
                          "requestedCount": 1,
                          "heldCount": 1,
                          "status": "SUCCESS"
                        }
                      ]
                    },
                    "message": "Hold created successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 픽업 시각 또는 영업시간 외",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    },
                    "message": {
                      "type": "string",
                      "example": "Pickup time must be a valid future datetime"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "사용자 또는 매장 없음",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    },
                    "message": {
                      "type": "string",
                      "example": "Store not found"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "재고 임시 hold 생성 (All-or-Nothing)",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/v1/reservations/confirm": {
      "post": {
        "description": "holdToken 기반으로 Redis Hold를 조회하고 DB에 예약을 확정합니다. 재고 차감은 이 단계에서 조건부 UPDATE(동시성 안전)로 수행됩니다. 확정 후 Redis Hold는 삭제됩니다.",
        "operationId": "confirmReservation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmHoldDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "예약 확정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "id": 1,
                      "userId": 1,
                      "status": "CONFIRMED",
                      "pickupTime": "2026-05-08T14:00:00.000Z",
                      "createdAt": "2026-05-07T10:00:00.000Z",
                      "items": [
                        {
                          "id": 1,
                          "inventoryId": 10,
                          "qty": 2
                        },
                        {
                          "id": 2,
                          "inventoryId": 20,
                          "qty": 1
                        }
                      ]
                    },
                    "message": "Reservation confirmed successfully"
                  }
                }
              }
            }
          },
          "404": {
            "description": "holdToken 없음 또는 만료",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    },
                    "message": {
                      "type": "string",
                      "example": "Hold not found or already expired"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "재고 부족 (confirm 시점 재확인)",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    },
                    "message": {
                      "type": "string",
                      "example": "Out of stock"
                    }
                  }
                }
              }
            }
          },
          "410": {
            "description": "hold TTL 만료",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    },
                    "message": {
                      "type": "string",
                      "example": "Hold token has expired"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Hold 확정 → 예약 생성",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/v1/reservations": {
      "get": {
        "description": "userId와 status 필터로 사용자의 예약 내역을 조회합니다. status는 confirmed(확정) 또는 cancelled(취소) 중 하나를 전달해야 합니다. 결과는 생성일 기준 최신순으로 반환됩니다.",
        "operationId": "listReservations",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "description": "조회할 사용자 ID",
            "schema": {
              "maximum": 9007199254740991,
              "exclusiveMinimum": 0,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "description": "예약 상태 필터 (confirmed: 확정 내역 | cancelled: 취소 내역)",
            "schema": {
              "enum": [
                "confirmed",
                "cancelled"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "예약 목록 반환 성공",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1,
                            "description": "예약 ID"
                          },
                          "userId": {
                            "type": "integer",
                            "example": 1,
                            "description": "사용자 ID"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "CONFIRMED",
                              "CANCELLED"
                            ],
                            "example": "CONFIRMED",
                            "description": "예약 상태 (CONFIRMED: 확정 | CANCELLED: 취소)"
                          },
                          "pickupTime": {
                            "type": "string",
                            "example": "2026-05-08T14:00:00.000Z",
                            "description": "픽업 예정 시각 (ISO 8601)"
                          },
                          "createdAt": {
                            "type": "string",
                            "example": "2026-05-07T10:00:00.000Z",
                            "description": "예약 생성 시각 (ISO 8601)"
                          },
                          "items": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "integer",
                                  "example": 1,
                                  "description": "예약 아이템 ID"
                                },
                                "inventoryId": {
                                  "type": "integer",
                                  "example": 10,
                                  "description": "재고 ID"
                                },
                                "qty": {
                                  "type": "integer",
                                  "example": 2,
                                  "description": "예약 수량"
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Reservations fetched successfully"
                    }
                  },
                  "example": {
                    "data": [
                      {
                        "id": 1,
                        "userId": 1,
                        "status": "CONFIRMED",
                        "pickupTime": "2026-05-08T14:00:00.000Z",
                        "createdAt": "2026-05-07T10:00:00.000Z",
                        "items": [
                          {
                            "id": 1,
                            "inventoryId": 10,
                            "qty": 2
                          },
                          {
                            "id": 2,
                            "inventoryId": 20,
                            "qty": 1
                          }
                        ]
                      }
                    ],
                    "message": "Reservations fetched successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 쿼리 파라미터 (userId 미입력, status 범위 오류 등)",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    },
                    "message": {
                      "type": "string",
                      "example": "Validation failed"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "예약 목록 조회",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/v1/reservations/{id}": {
      "get": {
        "operationId": "findReservation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "예약 정보",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "id": 1,
                      "userId": 1,
                      "status": "CONFIRMED",
                      "pickupTime": "2026-05-08T14:00:00.000Z",
                      "createdAt": "2026-05-07T10:00:00.000Z",
                      "items": [
                        {
                          "id": 1,
                          "inventoryId": 10,
                          "qty": 2
                        }
                      ]
                    },
                    "message": "Reservation fetched successfully"
                  }
                }
              }
            }
          },
          "404": {
            "description": "예약 없음",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    },
                    "message": {
                      "type": "string",
                      "example": "Reservation not found"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "예약 단건 조회",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/v1/reservations/{id}/cancel": {
      "post": {
        "description": "userId 권한 검증 후 CANCELLED 상태로 변경합니다. 픽업 1시간 미만 남은 경우 10% 취소 수수료 메시지를 반환합니다. 취소 시 재고가 자동으로 복구됩니다.",
        "operationId": "cancelReservation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelReservationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "취소 성공",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "id": 1,
                      "status": "CANCELLED",
                      "feeMessage": "전액 환불됩니다."
                    },
                    "message": "Reservation cancelled successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "이미 취소된 예약",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    },
                    "message": {
                      "type": "string",
                      "example": "Reservation already cancelled"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "예약 없음",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    },
                    "message": {
                      "type": "string",
                      "example": "Reservation not found"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "예약 취소",
        "tags": [
          "Reservations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bread API",
    "description": "Bread Reservation & Inventory API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateHoldDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "예약 사용자 ID"
          },
          "storeId": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "예약 대상 매장 ID"
          },
          "pickupTime": {
            "type": "string",
            "description": "픽업 예정 시각 (ISO 8601). 현재 시각 이후여야 하며 매장 영업시간(open_time ~ close_time) 범위 내여야 합니다."
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "breadId": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991,
                  "description": "빵 ID"
                },
                "qty": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 9007199254740991,
                  "description": "예약 수량 (1 이상)"
                }
              },
              "required": [
                "breadId",
                "qty"
              ]
            },
            "description": "예약할 빵 목록 (한 매장에서 여러 종류 동시 예약 가능)"
          }
        },
        "required": [
          "userId",
          "storeId",
          "pickupTime",
          "items"
        ]
      },
      "ConfirmHoldDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "예약 사용자 ID"
          },
          "holdToken": {
            "type": "string",
            "description": "hold 생성 시 발급된 토큰 (TTL 2분)"
          }
        },
        "required": [
          "userId",
          "holdToken"
        ]
      },
      "CancelReservationDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "취소 요청 사용자 ID (권한 검증)"
          }
        },
        "required": [
          "userId"
        ]
      },
...
    }
  }
}